+++ packages/shared/src/index.ts
@@ -1,5 +1,6 @@
 export * from './enums';
 export * from './constants';
+export * from './display';
 export * from './types';
 export * from './schemas';
 export * from './ws-messages';